requirements.txt
flask
PyPDF2
python-docx
nltk
spacy
